backend/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
numpy
groq
python-dotenv
python-jose
passlib[bcrypt]
python-multipart
sqlalchemy
pyodbc
langchain
langchain-groq
langchain-core
pypdf
python-docx
Pillow